+++ lib/sam_uiz.dart
@@ -1,7 +1,5 @@
 library sam_uiz;
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/dizzex_button.dart';
+export 'src/dizzex_card.dart';
+export 'src/dizzex_input_box.dart';