+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
 
-* TODO: Describe initial release.
+# Changelog
+
+## 0.0.1 - 2024-12-19
+- Initial release of `sam_uiz` package.
+- Added `DizzexButton`, `DizzexCard`, and `DizzexInputBox` components.